alert confirm sebelum hapus

+++ application/views/daftarunit.php
@@ -37,7 +37,7 @@
                         <td><?php echo $fetchdata['kode_unit']; ?></td>
                         <td><?php echo $fetchdata['nama_unit']; ?></td>
                         <td><a href="<?php echo base_url();?>index.php/unit/ubah/<?php echo $fetchdata['kode_unit'];?>">Edit</a> | 
-                            <a href="<?php echo base_url();?>index.php/unit/hapus/<?php echo $fetchdata['kode_unit'];?>">Hapus</a> | 
+                            <a href="<?php echo base_url();?>index.php/unit/hapus/<?php echo $fetchdata['kode_unit'];?>" onclick="return confirm_delete()">Hapus</a> | 
                             <a href="<?php echo base_url();?>index.php/unit/detail/<?php echo $fetchdata['kode_unit'];?>">Detail</a></td>
                       </tr>
                       <?php } ?>
@@ -57,3 +57,8 @@
           </div>
         </section>  
 </div>        
+<script type="text/javascript">
+function confirm_delete() {
+  return confirm('apa anda yakin ingin menghapus ?');
+}
+</script>
